% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_shiny.R
\name{pg_demo}
\alias{pg_demo}
\title{Run Web-Based Demonstration}
\usage{
pg_demo()
}
\description{
Call Shiny to run a web-based application for the demonstration of entropy
balancing
}
\examples{
if(interactive()){
pg_demo()}

}
