% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_entropy.R
\name{rwe_extract_stats_covx}
\alias{rwe_extract_stats_covx}
\title{Get estimating equation}
\usage{
rwe_extract_stats_covx(target_stats, dta, cont_stat = c("mean", "ex2"))
}
\arguments{
\item{target_stats}{Summary statistics to be extracted}

\item{dta}{Data frame with subject level data}
}
\value{
List of summary statistics
}
\description{
Extract summary statistics from a data frame based on existing statistics
}
\examples{
rwe_extract_stats(ex_stats, ex_rwd)

}
