% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_entropy.R
\name{rwe_margin_entropy}
\alias{rwe_margin_entropy}
\title{Calculate entropy balancing score}
\usage{
rwe_margin_entropy(
  target_stats,
  dta_ext,
  tol = 1e-08,
  max_it = 10000,
  print_level = 0
)
}
\arguments{
\item{target_stats}{List of summary statistics in a target patient population}

\item{dta_ext}{External data dataframe}

\item{tol}{Tolerance level for entropy balancing optimization}

\item{max_it}{Maximum number of iterations}

\item{print_level}{Verbose or not}
}
\value{
A list that includes weights (entropy balancing weights), stats
    (summary statistics based on weighted patients), convergence status, etc.
}
\description{
Calculate entropy balancing score
}
\examples{
rwe_margin_entropy(ex_stats, ex_rwd)

}
