% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_entropy.R
\name{rwe_pgoal}
\alias{rwe_pgoal}
\title{Estimate Performance Goal}
\usage{
rwe_pgoal(
  target_stats,
  exist_stats,
  exist_ey,
  dta_ext,
  v_y = "Y",
  nbs = 100,
  ...
)
}
\arguments{
\item{target_stats}{List of summary statistics in a target patient population}

\item{exist_stats}{List of summary statistics in an existing clinical study}

\item{exist_ey}{Expectation of the outcome reported by the existing clinical study}

\item{dta_ext}{External data dataframe}

\item{v_y}{Column name in the external data dataframe corresponding to
outcome}

\item{nbs}{Number of bootstraps}

\item{...}{Optional parameters for entropy balancing}
}
\description{
Different estimates of the outcomes for deriving performance goals.
}
\examples{
rst <- rwe_pgoal(target_stats = ex_stats,
                 exist_stats  = ex2_stats,
                 exist_ey     = 0.1,
                 dta_ext      = ex_rwd,
                 v_y          = "dead_6mths")
print(rst$est_pg)


}
