% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_entropy.R
\name{rwe_pgoal_est}
\alias{rwe_pgoal_est}
\title{Estimate Adjusted Mean of Outcome}
\usage{
rwe_pgoal_est(
  target_stats,
  dta_ext,
  v_y = "Y",
  reported = NULL,
  nbs = 100,
  ...
)
}
\arguments{
\item{target_stats}{List of summary statistics in a target patient population}

\item{dta_ext}{External data dataframe}

\item{v_y}{Column name in the external data dataframe corresponding to
outcome}

\item{reported}{Reported mean outcome of the target patient population}

\item{nbs}{Number of bootstraps}

\item{...}{Optional parameters for entropy balancing}
}
\description{
Estimate adjusted mean of outcome by entropy balancing and operation
difference if mean outcome was reported in a clinical study
}
\examples{
rwe_pgoal_est(ex_stats, ex_rwd, v_y = "dead_6mths",  nbs = 10)

}
